Arrow code style
There's a good reason to place -> and => on the same line as the
preceding argument, not as the following one: -> and => change the
semantics of the preceding argument, `(v : Ty) =>` really meaning
`{auto 0 v : Ty}`.
This means  it's convenient to have the specific arrow on the same line
as the argument whose semantics it potentially modifies.